requirements.txt
MouseInfo==0.1.3
Pillow==8.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dotenv==0.19.2
pytweening==1.0.4
pywin32==302
schedule==1.1.0
